+++ src/commands/build-preact-app.js
@@ -37,8 +37,8 @@ function buildConfig(args) {
     },
     resolve: {
       alias: {
-        'react': 'preact-compat',
-        'react-dom': 'preact-compat',
+        'react': 'preact-compat/dist/preact-compat',
+        'react-dom': 'preact-compat/dist/preact-compat',
       }
     },
   }