+++ src/commands/build-preact.js
@@ -41,8 +41,8 @@ function buildConfig(args) {
     },
     resolve: {
       alias: {
-        'react': 'preact-compat',
-        'react-dom': 'preact-compat',
+        'react': 'preact-compat/dist/preact-compat',
+        'react-dom': 'preact-compat/dist/preact-compat',
       }
     },
   }