+++ src/createServePreactAppConfig.js
@@ -24,8 +24,8 @@ export default function createServePreactAppConfig(args, overrides) {
     },
     resolve: {
       alias: {
-        'react': 'preact-compat',
-        'react-dom': 'preact-compat',
+        'react': 'preact-compat/dist/preact-compat',
+        'react-dom': 'preact-compat/dist/preact-compat',
       }
     },
   }